+++ lib/patch_util/git.rb
@@ -3,6 +3,7 @@
 module PatchUtil
   module Git
     autoload :Cli, 'patch_util/git/cli'
+    autoload :RewritePreflightVerifier, 'patch_util/git/rewrite_preflight_verifier'
     autoload :RewriteStateStore, 'patch_util/git/rewrite_state_store'
     autoload :RewriteSessionManager, 'patch_util/git/rewrite_session_manager'
     autoload :Rewriter, 'patch_util/git/rewriter'